+++ libretro-common/include/streams/interface_stream.h
@@ -48,7 +48,7 @@ typedef struct intfstream_info
       struct
       {
          uint8_t *data;
-         unsigned size;
+         uint64_t size;
       } buf;
       bool writable;
    } memory;
